+++ src/main/java/com/cursos/repository/CursosRepository.java
@@ -8,5 +8,5 @@ import com.cursos.model.Curso;
 
 public interface CursosRepository extends JpaRepository<Curso, Integer> {
     @Query("SELECT c FROM Curso c WHERE c.precio BETWEEN :priceMin AND :priceMax")
-    List<Curso> getPreciosBetween(int priceMin, int priceMax);
+    List<Curso> getPreciosBetween(double priceMin, double priceMax);
 }
